# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/quiz/questions:
    get:
      description: OWNER, ADMIN, and STAFF can read paginated quiz questions.
      operationId: AdminQuizController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: coverage
            type: string
        - name: attributeGroupId
          required: false
          in: query
          schema:
            example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
            type: string
        - name: selectionType
          required: false
          in: query
          schema:
            type: string
            enum:
              - SINGLE
              - MULTIPLE
        - name: isRequired
          required: false
          in: query
          schema:
            example: true
            type: boolean
        - name: isActive
          required: false
          in: query
          schema:
            example: true
            type: boolean
        - name: sortBy
          required: false
          in: query
          schema:
            type: string
            enum:
              - stepOrder
              - createdAt
              - questionText
        - name: sortOrder
          required: false
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        "200":
          description: Paginated quiz questions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestionResponse"
        "401": &a16
          description: Authentication required.
        "403": &a17
          description: Insufficient admin role.
      security: &a18
        - bearer: []
      summary: List admin quiz questions
      tags: &a19
        - Admin Quiz
    post:
      description: OWNER and ADMIN can create quiz questions and mapped answer
        options. Only one active question is allowed per attribute group.
      operationId: AdminQuizController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizQuestionDto"
      responses:
        "200":
          description: Quiz question created.
        "401": *a16
        "403": *a17
        "409":
          description: Only one active question is allowed per attribute group.
      security: *a18
      summary: Create quiz question
      tags: *a19
  /admin/quiz/questions/reorder:
    patch:
      description: OWNER and ADMIN can update quiz question step order in bulk.
      operationId: AdminQuizController_reorder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderQuizQuestionsDto"
      responses:
        "200":
          description: Quiz questions reordered.
        "401": *a16
        "403": *a17
      security: *a18
      summary: Reorder quiz questions
      tags: *a19
  /admin/quiz/questions/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can read quiz question details.
      operationId: AdminQuizController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Quiz question details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestionResponse"
        "401": *a16
        "403": *a17
        "404":
          description: Quiz question not found.
      security: *a18
      summary: Get admin quiz question details
      tags: *a19
    patch:
      description: OWNER and ADMIN can update quiz question fields and replace
        options. Active uniqueness per attribute group is enforced.
      operationId: AdminQuizController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuizQuestionDto"
      responses:
        "200":
          description: Quiz question updated.
        "401": *a16
        "403": *a17
        "404":
          description: Quiz question not found.
        "409":
          description: Only one active question is allowed per attribute group.
      security: *a18
      summary: Update quiz question
      tags: *a19
    delete:
      description: OWNER and ADMIN can soft-deactivate quiz questions.
      operationId: AdminQuizController_deactivate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Quiz question deactivated.
        "401": *a16
        "403": *a17
        "404":
          description: Quiz question not found.
      security: *a18
      summary: Deactivate quiz question
      tags: *a19
...
  /products/{id}:
    get:
      description: Returns one active product with category, brand, references, and
        compatibility attributes.
      operationId: ProductsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Product ID.
          schema:
            example: 00000000-0000-4000-8000-000000000001
            type: string
      responses:
        "200":
          description: Product details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Product not found.
      summary: Get an active product by ID
      tags: *a20
  /admin/products:
    get:
      description: OWNER, ADMIN, and STAFF can read paginated products with admin-only
        stock/count summaries.
      operationId: AdminProductsController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: foundation
            type: string
        - name: categoryId
          required: false
          in: query
          schema:
            example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
            type: string
        - name: brandId
          required: false
          in: query
          schema:
            example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - DRAFT
              - ACTIVE
              - ARCHIVED
        - name: isActive
          required: false
          in: query
          schema:
            example: true
            type: boolean
        - name: sortBy
          required: false
          in: query
          schema:
            type: string
            enum:
              - createdAt
              - name
              - basePrice
        - name: sortOrder
          required: false
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        "200":
          description: Paginated products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401": &a21
          description: Authentication required.
        "403": &a22
          description: Insufficient admin role.
      security: &a23
        - bearer: []
      summary: List admin products
      tags: &a24
        - Admin Products
    post:
      description: OWNER and ADMIN can create catalog products.
      operationId: AdminProductsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "200":
          description: Product created.
        "401": *a21
        "403": *a22
        "409":
          description: Product slug already exists.
      security: *a23
      summary: Create product
      tags: *a24
  /admin/products/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can read full product details.
      operationId: AdminProductsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Product details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401": *a21
        "403": *a22
        "404":
          description: Product not found.
      security: *a23
      summary: Get admin product details
      tags: *a24
    patch:
      description: OWNER and ADMIN can update mutable product fields.
      operationId: AdminProductsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDto"
      responses:
        "200":
          description: Product updated.
        "401": *a21
        "403": *a22
        "404":
          description: Product not found.
        "409":
          description: Product slug already exists.
      security: *a23
      summary: Update product
      tags: *a24
    delete:
      description: OWNER and ADMIN can archive a product and deactivate its references.
      operationId: AdminProductsController_archive
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Product archived.
        "401": *a21
        "403": *a22
        "404":
          description: Product not found.
      security: *a23
      summary: Archive product
      tags: *a24
...
info:
  title: Beauty Pack Recommendation API
  description: Backend API for beauty-pack personalization, recommendations, COD
    orders, and administration
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the JWT access token returned by POST /auth/login.
  schemas:
...
    AttributeOptionResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: MEDIUM
        label:
          type: string
          example: Medium
      required:
        - id
        - code
        - label
...
    QuizAttributeGroupResponse:
      type: object
      properties:
        code:
          type: string
          example: SKIN_COLOR
        name:
          type: string
          example: Skin Color
      required:
        - code
        - name
    QuizQuestionOptionResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        attributeOptionId:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: MEDIUM
        label:
          type: string
          example: Medium
        displayLabel:
          type: object
          example: null
          nullable: true
        displayImageUrl:
          type: object
          example: null
          nullable: true
      required:
        - id
        - attributeOptionId
        - code
        - label
    QuizQuestionResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        questionText:
          type: string
          example: What is your skin color?
        helperText:
          type: object
          example: null
          nullable: true
        selectionType:
          type: string
          enum:
            - SINGLE
            - MULTIPLE
          example: SINGLE
        isRequired:
          type: boolean
          example: true
        stepOrder:
          type: number
          example: 1
        attributeGroup:
          $ref: "#/components/schemas/QuizAttributeGroupResponse"
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionOptionResponse"
      required:
        - id
        - questionText
        - selectionType
        - isRequired
        - stepOrder
        - attributeGroup
        - options
...
    ReorderQuizQuestionItemDto:
      type: object
      properties:
        questionId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        stepOrder:
          type: number
          example: 1
      required:
        - questionId
        - stepOrder
    ReorderQuizQuestionsDto:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ReorderQuizQuestionItemDto"
      required:
        - questions
    QuizQuestionOptionInputDto:
      type: object
      properties:
        attributeOptionId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        displayLabel:
          type: object
          example: Medium
        displayImageUrl:
          type: object
          example: https://example.com/medium.jpg
        sortOrder:
          type: number
          example: 2
          default: 0
        isActive:
          type: boolean
          example: true
          default: true
      required:
        - attributeOptionId
    CreateQuizQuestionDto:
      type: object
      properties:
        attributeGroupId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        questionText:
          type: string
          example: What coverage do you prefer?
        helperText:
          type: object
          example: Choose your preferred result
        selectionType:
          type: string
          enum: &a45
            - SINGLE
            - MULTIPLE
          default: SINGLE
        isRequired:
          type: boolean
          example: true
          default: true
        stepOrder:
          type: number
          example: 6
          default: 0
        isActive:
          type: boolean
          example: false
          default: false
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionOptionInputDto"
      required:
        - attributeGroupId
        - questionText
    UpdateQuizQuestionDto:
      type: object
      properties:
        attributeGroupId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        questionText:
          type: string
          example: What coverage do you prefer?
        helperText:
          type: object
          example: Choose your preferred result
        selectionType:
          type: string
          enum: *a45
          default: SINGLE
        isRequired:
          type: boolean
          example: true
          default: true
        stepOrder:
          type: number
          example: 6
          default: 0
        isActive:
          type: boolean
          example: false
          default: false
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionOptionInputDto"
    CategorySummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: FOUNDATION
        name:
          type: string
          example: Foundation
      required:
        - id
        - code
        - name
    BrandSummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Demo Beauty
      required:
        - id
        - name
    AttributeMatchResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        matchType:
          type: string
          enum:
            - COMPATIBLE
            - NOT_COMPATIBLE
            - BOOST
          example: COMPATIBLE
        scoreValue:
          type: number
          example: 40
        isHardFilter:
          type: boolean
          example: false
        attributeGroup:
          $ref: "#/components/schemas/QuizAttributeGroupResponse"
        attributeOption:
          $ref: "#/components/schemas/AttributeOptionResponse"
      required:
        - id
        - matchType
        - scoreValue
        - isHardFilter
        - attributeGroup
        - attributeOption
    ProductReferenceResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        referenceCode:
          type: string
          example: RF2
        referenceName:
          type: string
          example: Medium Warm
        sku:
          type: object
          example: FOUNDATION-X-RF2
          nullable: true
        barcode:
          type: object
          example: null
          nullable: true
        priceDelta:
          type: number
          example: 0
        stockQuantity:
          type: number
          example: 20
        reservedQuantity:
          type: number
          example: 0
        isActive:
          type: boolean
          example: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/AttributeMatchResponse"
      required:
        - id
        - referenceCode
        - referenceName
        - priceDelta
        - stockQuantity
        - reservedQuantity
        - isActive
        - attributes
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Foundation X
        slug:
          type: string
          example: foundation-x
        basePrice:
          type: number
          example: 120
        currency:
          type: string
          example: MAD
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - ARCHIVED
          example: ACTIVE
        category:
          $ref: "#/components/schemas/CategorySummaryResponse"
        brand:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/BrandSummaryResponse"
        references:
          type: array
          items:
            $ref: "#/components/schemas/ProductReferenceResponse"
      required:
        - id
        - name
        - slug
        - basePrice
        - currency
        - status
        - category
        - references
    CreateProductDto:
      type: object
      properties:
        categoryId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        brandId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        name:
          type: string
          example: Foundation X
        slug:
          type: string
          example: foundation-x
        description:
          type: string
          example: Demo foundation
        basePrice:
          type: number
          example: 120
        costPrice:
          type: number
          example: 70
        currency:
          type: string
          example: MAD
          default: MAD
        mainImageUrl:
          type: string
          example: https://example.com/product.jpg
        status:
          type: string
          enum: &a46
            - DRAFT
            - ACTIVE
            - ARCHIVED
          default: DRAFT
        isActive:
          type: boolean
          example: true
          default: true
      required:
        - categoryId
        - name
        - slug
        - basePrice
        - currency
    UpdateProductDto:
      type: object
      properties:
        categoryId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        brandId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        name:
          type: string
          example: Foundation X
        slug:
          type: string
          example: foundation-x
        description:
          type: string
          example: Demo foundation
        basePrice:
          type: number
          example: 120
        costPrice:
          type: number
          example: 70
        currency:
          type: string
          example: MAD
          default: MAD
        mainImageUrl:
          type: string
          example: https://example.com/product.jpg
        status:
          type: string
          enum: *a46
          default: DRAFT
        isActive:
          type: boolean
          example: true
          default: true
...